requirements.txt
requests
PyJWT
flask
pre-commit
